+++ packages/atomic/src/components/facets/atomic-facet/atomic-facet.tsx
@@ -153,7 +153,7 @@ export class AtomicFacet
   // @Prop() public customSort?: string; TODO: KIT-753 Add customSort option for facet
 
   /**
-   * The required facets & values for this facet to be displayed.
+   * The required facets and values for this facet to be displayed.
    * Examples:
    * ```html
    * <atomic-facet facet-id="abc" field="objecttype" ...></atomic-facet>
@@ -171,7 +171,7 @@ export class AtomicFacet
    * ></atomic-facet>
    * ```
    */
-  @MapProp() public dependsOn: Record<string, string> = {};
+  @MapProp() @Prop() public dependsOn: Record<string, string> = {};
 
   @FocusTarget()
   private showMoreFocus!: FocusTargetController;