+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.15.0...@coveo/quantic@1.16.0) (2022-05-13)
+
+
+### Features
+
+* **quantic:** Create QuanticRefineModalContent component ([#1948](https://github.com/coveo/ui-kit/issues/1948)) ([97dca15](https://github.com/coveo/ui-kit/commit/97dca1547c17e72510d8cee34cb7f7cab217c9c4))
+
+
+
+
+
+# [1.15.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.14.2...@coveo/quantic@1.15.0) (2022-05-11)
+
+
+### Features
+
+* **atomic:** network improvements ([#1981](https://github.com/coveo/ui-kit/issues/1981)) ([6490e0d](https://github.com/coveo/ui-kit/commit/6490e0dd0ca16e28e3aa715b87ed2cdeb66047f1))
+
+
+
+
+
+## [1.14.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.14.1...@coveo/quantic@1.14.2) (2022-05-11)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [1.14.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.14.0...@coveo/quantic@1.14.1) (2022-05-09)
 
 **Note:** Version bump only for package @coveo/quantic